db/connection.py
import os
import psycopg2
from sqlalchemy import create_url

def get_db_connection():
    """
    Returns a connection to the PostgreSQL database.
    Placeholder for future connection management.
    """
    db_url = os.environ.get("DATABASE_URL")
    # For now, just return None or mocked connection
    return None
