requirements.txt
flask==3.0.2
flask-cors==4.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27
python-dotenv==1.0.1
gunicorn==21.2.0
structlog==24.1.0
python-json-logger==2.0.7
requests==2.31.0
